combined HPA test files
